Finished Fortress Runes, Fire Rate Tweak, Organisation
Fortress Runes have been designed and implemented. Each element has a path that branches off at the end and only one can be chosen. The upgrades apply to all placed towers and to towers placed after purchase. The other two paths upgrade the rate and amount of mana gained. Values used for upgrades and costs are provisional, but aren't intended to be largely changed in the future.

Tower runes that provide fire rate now provide a flat increase (previously percentage) and tower types now have varied fire rates.

A lot of variables were made static as the scripts they are contained in are in scripts that are attached to one object.

Made prefab of fortress and moved existing prefabs into the blueprint folder.