Tower Range Indicator
Clicking on a tower (i.e. opening the tower menu) now shows a gray-white circle indicating the tower's range.

If a tower rune is bought and it alters range, the circle updates to reflect this.

Adjusted tower prefabs so that the range is calculated from the floor, and the towers now also sit on top of the pedestals.